+++ src/components/common/bottomSheet/BottomSheetStyle.tsx
@@ -7,6 +7,7 @@ export const Background = styled.div`
   width: 100%;
   height: 100%;
   background-color: rgba(0, 0, 0, 0.5);
+  z-index: 100;
 `;
 
 export const BottomSheet = styled.div<{ $type: string }>`
@@ -18,6 +19,7 @@ export const BottomSheet = styled.div<{ $type: string }>`
   padding: 1.25rem;
   border-radius: 1.25rem 1.25rem 0rem 0rem;
   background-color: ${({ theme }) => theme.colors.white};
+  z-index: 100px;
 `;
 
 export const Header = styled.header`
