+++ src/components/common/bottomSheet/RecordBottomSheetStyle.tsx
